refactor: centralize theming and add bottom navigation bar
