+++ lib/app/app.dart
@@ -1,17 +1,11 @@
+export 'app_config.dart' show AppConfig, AppConfigData;
 export 'data.dart';
 export 'hive.dart';
 export 'services/network.dart';
 export 'services/storage.dart';
 export 'services/user_fetcher.dart';
-export 'theming/brb.dart' show brbAppConfig;
-export 'theming/config.dart' show AppConfigData, AppConfig;
-export 'theming/n21.dart' show n21AppConfig;
-export 'theming/open.dart' show openAppConfig;
-export 'theming/sc.dart' show schulCloudAppConfig;
-export 'theming/thr.dart' show thrAppConfig;
-export 'theming/utils.dart';
+export 'theming_utils.dart';
 export 'utils.dart';
-export 'widgets/app_bar.dart' show AppBarActions;
 export 'widgets/buttons.dart';
 export 'widgets/empty_state.dart';
 export 'widgets/error_widgets.dart';
