+++ lib/app/widgets/app_bar.dart
@@ -1,19 +1,15 @@
-import 'dart:ui';
-
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:schulcloud/app/app.dart';
+import 'package:schulcloud/l10n/l10n.dart';
 
-import 'menu.dart';
 import 'schulcloud_app.dart';
 
 final _appBarKey = GlobalKey();
 
 /// A custom version of an app bar intended to be displayed at the bottom of
-/// the screen. You can also also call [_MyAppBarState.register] and
-/// [_MyAppBarState.unregister] to register and unregister actions on the app
-/// bar. The [AppBarActions] does that.
-class MyAppBar extends StatefulWidget {
+/// the screen.
+class MyAppBar extends StatelessWidget {
   MyAppBar({
     @required this.onNavigate,
     @required this.activeScreenStream,
@@ -24,128 +20,40 @@ class MyAppBar extends StatefulWidget {
   final Stream<Screen> activeScreenStream;
 
   @override
-  _MyAppBarState createState() => _MyAppBarState();
-}
-
-class _MyAppBarState extends State<MyAppBar> {
-  final _actionsByState = <State<AppBarActions>, List<Widget>>{};
-  final _actions = <Widget>[];
-
-  void register(State<AppBarActions> state, List<Widget> actions) {
-    Future.microtask(() {
-      setState(() {
-        _actions.addAll(actions);
-        _actionsByState[state] = actions;
-      });
-    });
-  }
-
-  void unregister(State<AppBarActions> state) {
-    Future.microtask(() {
-      setState(() {
-        final actionsToRemove = _actionsByState.remove(state);
-        _actions.removeWhere(actionsToRemove.contains);
-      });
-    });
-  }
-
-  Future<void> _showMenu(BuildContext context) async {
-    final target = await context.navigator.push(PageRouteBuilder(
-      pageBuilder: (_, __, ___) =>
-          Menu(activeScreenStream: widget.activeScreenStream),
-      opaque: false,
-      maintainState: true,
-      transitionsBuilder: _customDialogTransitionBuilder,
-    ));
-    if (target != null) {
-      widget.onNavigate(target);
-    }
-  }
-
-  @override
   Widget build(BuildContext context) {
-    return Material(
-      color: context.theme.bottomAppBarColor,
-      elevation: 12,
-      child: Container(
-        height: 56,
-        padding: EdgeInsets.symmetric(horizontal: 8),
-        alignment: Alignment.center,
-        child: IconTheme(
-          data: IconThemeData(color: context.theme.contrastColor),
-          child: Row(
-            children: <Widget>[
-              IconButton(
-                icon: Icon(Icons.menu),
-                onPressed: () => _showMenu(context),
-              ),
-              Spacer(),
-              ..._actions,
-            ],
-          ),
+    final s = context.s;
+
+    return BottomNavigationBar(
+      selectedItemColor: context.theme.primaryColor,
+      unselectedItemColor: context.theme.mediumEmphasisColor,
+      items: [
+        BottomNavigationBarItem(
+          icon: Icon(Icons.dashboard),
+          title: Text(s.app_navigation_dashboard),
         ),
-      ),
+        BottomNavigationBarItem(
+          icon: Icon(Icons.new_releases),
+          title: Text(s.app_navigation_news),
+        ),
+        BottomNavigationBarItem(
+          icon: Icon(Icons.school),
+          title: Text(s.app_navigation_courses),
+        ),
+        BottomNavigationBarItem(
+          icon: Icon(Icons.playlist_add_check),
+          title: Text(s.app_navigation_assignments),
+        ),
+        BottomNavigationBarItem(
+          icon: Icon(Icons.folder),
+          title: Text(s.app_navigation_files),
+        ),
+      ],
     );
+    /*buildItem(Screen.dashboard, s.app_navigation_dashboard, Icons.dashboard),
+      buildItem(Screen.news, s.app_navigation_news, Icons.new_releases),
+      buildItem(Screen.courses, s.app_navigation_courses, Icons.school),
+      buildItem(Screen.homework, s.app_navigation_assignments,
+          Icons.playlist_add_check),
+      buildItem(Screen.files, s.app_navigation_files, Icons.folder),*/
   }
-}
-
-/// A widget that adds actions to the enclosing [MyAppBar].
-class AppBarActions extends StatefulWidget {
-  const AppBarActions({@required this.actions, @required this.child})
-      : assert(actions != null),
-        assert(child != null);
-
-  final List<Widget> actions;
-  final Widget child;
-
-  @override
-  _AppBarActionsState createState() => _AppBarActionsState();
-}
-
-class _AppBarActionsState extends State<AppBarActions> {
-  _MyAppBarState _findEnclosingMyAppBar() => _appBarKey.currentState;
-
-  @override
-  void initState() {
-    super.initState();
-    _findEnclosingMyAppBar().register(this, widget.actions);
-  }
-
-  @override
-  void dispose() {
-    _findEnclosingMyAppBar().unregister(this);
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) => widget.child;
-}
-
-Widget _customDialogTransitionBuilder(
-  BuildContext context,
-  Animation<double> animation,
-  Animation<double> secondaryAnimation,
-  Widget child,
-) {
-  return Stack(
-    children: <Widget>[
-      FadeTransition(
-        opacity: animation,
-        child: GestureDetector(
-          onTap: () => context.navigator.popUntil((route) => route.isFirst),
-          child: Container(color: Colors.black45),
-        ),
-      ),
-      SlideTransition(
-        position: Tween<Offset>(
-          begin: Offset(0, 1),
-          end: Offset.zero,
-        ).animate(CurvedAnimation(
-          parent: animation,
-          curve: Curves.easeOutQuad,
-        )),
-        child: Align(alignment: Alignment.bottomCenter, child: child),
-      ),
-    ],
-  );
 }
