+++ lib/l10n/l10n.dart
@@ -1,0 +1,2 @@
+export 'package:schulcloud/generated/l10n.dart';
+export 'package:schulcloud/generated/generated.dart';
